refactor message handlers
